requirements.txt
flask==1.1.4
python-dotenv==0.19.0
requests==2.26.0
waitress==2.0.0
werkzeug==0.16.1